refactor(automation): colocate the automation helpers and refresh on toggle
Two things from review.

The lifecycle labels and the output summariser were put in common/constants and
common/utils, where they read as generic abstractions the whole client can
build on. They are neither: both exist to describe an automation. The
architecture guide is explicit that common is for genuinely cross-feature code
and that feature-specific helpers stay with their owner even when another file
uses them, so both move next to the automations panel. Their one outside caller
is the event editor's trigger list, which is the same feature seen from the
rundown, and which already reaches into app-settings for its navigation helper.
common/constants is gone with them: it held nothing else, because this branch
invented it.

The automations list keeps its original wording for the disabled notice. What
was wrong was underneath it: the settings form saved the master switch and
never refetched, so every other part of the panel kept the stale answer until
the slow poll came round and turning automations on appeared to do nothing.
Every other form in this panel refetches after saving; this one now does too.

Verified in the running app: with automations off and one automation in the
list, flipping the switch and saving clears the notice in about 250ms rather
than waiting out the poll.
