+++ apps/client/src/features/app-settings/panel/automations-panel/__tests__/automationOutputs.test.ts
