+++ apps/client/src/features/app-settings/panel/automations-panel/automationOutputs.ts
