+++ apps/client/src/features/app-settings/panel/automations-panel/timerLifecycle.ts
